Backend finalizado pero con algunos errores

+++ src/main/java/com/sevilla/clubtor/controllers/ClubController.java
@@ -1,6 +1,7 @@
 package com.sevilla.clubtor.controllers;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import com.sevilla.clubtor.models.ClubModel;
@@ -22,8 +23,13 @@ public class ClubController {
     }
 
     @GetMapping("/{id}")
-    public Optional<ClubModel> getClubById(@PathVariable(value = "id") Long clubId) {
-        return clubService.getClubById(clubId);
+    public ResponseEntity<ClubModel> getClubById(@PathVariable(value = "id") Long clubId) {
+        Optional<ClubModel> club = clubService.getClubById(clubId);
+        if (club.isPresent()) {
+            return ResponseEntity.ok().body(club.get());
+        } else {
+            return ResponseEntity.notFound().build();
+        }
     }
 
     @PostMapping
@@ -32,13 +38,24 @@ public class ClubController {
     }
 
     @PutMapping("/{id}")
-    public ClubModel updateClub(@PathVariable(value = "id") Long clubId, @RequestBody ClubModel clubDetails) {
-        return clubService.updateClub(clubId, clubDetails);
+    public ResponseEntity<ClubModel> updateClub(@PathVariable(value = "id") Long clubId, @RequestBody ClubModel clubDetails) {
+        Optional<ClubModel> club = clubService.getClubById(clubId);
+        if (club.isPresent()) {
+            ClubModel updatedClub = clubService.updateClub(clubId, clubDetails);
+            return ResponseEntity.ok(updatedClub);
+        } else {
+            return ResponseEntity.notFound().build();
+        }
     }
 
     @DeleteMapping("/{id}")
-    public void deleteClub(@PathVariable(value = "id") Long clubId) {
-        clubService.deleteClub(clubId);
+    public ResponseEntity<Void> deleteClub(@PathVariable(value = "id") Long clubId) {
+        Optional<ClubModel> club = clubService.getClubById(clubId);
+        if (club.isPresent()) {
+            clubService.deleteClub(clubId);
+            return ResponseEntity.noContent().build();
+        } else {
+            return ResponseEntity.notFound().build();
+        }
     }
 }
-
