+++ src/main/java/com/sevilla/clubtor/controllers/EventController.java
@@ -1,6 +1,7 @@
 package com.sevilla.clubtor.controllers;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import com.sevilla.clubtor.models.EventModel;
@@ -22,8 +23,13 @@ public class EventController {
     }
 
     @GetMapping("/{id}")
-    public Optional<EventModel> getEventById(@PathVariable(value = "id") Long eventId) {
-        return eventService.getEventById(eventId);
+    public ResponseEntity<EventModel> getEventById(@PathVariable(value = "id") Long eventId) {
+        Optional<EventModel> event = eventService.getEventById(eventId);
+        if (event.isPresent()) {
+            return ResponseEntity.ok().body(event.get());
+        } else {
+            return ResponseEntity.notFound().build();
+        }
     }
 
     @PostMapping
@@ -32,13 +38,24 @@ public class EventController {
     }
 
     @PutMapping("/{id}")
-    public EventModel updateEvent(@PathVariable(value = "id") Long eventId, @RequestBody EventModel eventDetails) {
-        return eventService.updateEvent(eventId, eventDetails);
+    public ResponseEntity<EventModel> updateEvent(@PathVariable(value = "id") Long eventId, @RequestBody EventModel eventDetails) {
+        Optional<EventModel> event = eventService.getEventById(eventId);
+        if (event.isPresent()) {
+            EventModel updatedEvent = eventService.updateEvent(eventId, eventDetails);
+            return ResponseEntity.ok(updatedEvent);
+        } else {
+            return ResponseEntity.notFound().build();
+        }
     }
 
     @DeleteMapping("/{id}")
-    public void deleteEvent(@PathVariable(value = "id") Long eventId) {
-        eventService.deleteEvent(eventId);
+    public ResponseEntity<Void> deleteEvent(@PathVariable(value = "id") Long eventId) {
+        Optional<EventModel> event = eventService.getEventById(eventId);
+        if (event.isPresent()) {
+            eventService.deleteEvent(eventId);
+            return ResponseEntity.noContent().build();
+        } else {
+            return ResponseEntity.notFound().build();
+        }
     }
 }
-
