+++ src/main/java/com/sevilla/clubtor/controllers/MatchController.java
@@ -1,6 +1,7 @@
 package com.sevilla.clubtor.controllers;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import com.sevilla.clubtor.models.MatchModel;
@@ -22,8 +23,13 @@ public class MatchController {
     }
 
     @GetMapping("/{id}")
-    public Optional<MatchModel> getMatchById(@PathVariable(value = "id") Long matchId) {
-        return matchService.getMatchById(matchId);
+    public ResponseEntity<MatchModel> getMatchById(@PathVariable(value = "id") Long matchId) {
+        Optional<MatchModel> match = matchService.getMatchById(matchId);
+        if (match.isPresent()) {
+            return ResponseEntity.ok().body(match.get());
+        } else {
+            return ResponseEntity.notFound().build();
+        }
     }
 
     @PostMapping
@@ -32,12 +38,25 @@ public class MatchController {
     }
 
     @PutMapping("/{id}")
-    public MatchModel updateMatch(@PathVariable(value = "id") Long matchId, @RequestBody MatchModel matchDetails) {
-        return matchService.updateMatch(matchId, matchDetails);
+    public ResponseEntity<MatchModel> updateMatch(@PathVariable(value = "id") Long matchId, @RequestBody MatchModel matchDetails) {
+        Optional<MatchModel> match = matchService.getMatchById(matchId);
+        if (match.isPresent()) {
+            MatchModel updatedMatch = matchService.updateMatch(matchId, matchDetails);
+            return ResponseEntity.ok(updatedMatch);
+        } else {
+            return ResponseEntity.notFound().build();
+        }
     }
 
     @DeleteMapping("/{id}")
-    public void deleteMatch(@PathVariable(value = "id") Long matchId) {
-        matchService.deleteMatch(matchId);
+    public ResponseEntity<Void> deleteMatch(@PathVariable(value = "id") Long matchId) {
+        Optional<MatchModel> match = matchService.getMatchById(matchId);
+        if (match.isPresent()) {
+            matchService.deleteMatch(matchId);
+            return ResponseEntity.noContent().build();
+        } else {
+            return ResponseEntity.notFound().build();
+        }
     }
 }
+
