+++ src/main/java/com/sevilla/clubtor/controllers/PlayerController.java
@@ -1,6 +1,7 @@
 package com.sevilla.clubtor.controllers;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import com.sevilla.clubtor.models.PlayerModel;
@@ -22,8 +23,13 @@ public class PlayerController {
     }
 
     @GetMapping("/{id}")
-    public Optional<PlayerModel> getPlayerById(@PathVariable(value = "id") Long playerId) {
-        return playerService.getPlayerById(playerId);
+    public ResponseEntity<PlayerModel> getPlayerById(@PathVariable(value = "id") Long playerId) {
+        Optional<PlayerModel> player = playerService.getPlayerById(playerId);
+        if (player.isPresent()) {
+            return ResponseEntity.ok().body(player.get());
+        } else {
+            return ResponseEntity.notFound().build();
+        }
     }
 
     @PostMapping
@@ -32,12 +38,24 @@ public class PlayerController {
     }
 
     @PutMapping("/{id}")
-    public PlayerModel updatePlayer(@PathVariable(value = "id") Long playerId, @RequestBody PlayerModel playerDetails) {
-        return playerService.updatePlayer(playerId, playerDetails);
+    public ResponseEntity<PlayerModel> updatePlayer(@PathVariable(value = "id") Long playerId, @RequestBody PlayerModel playerDetails) {
+        Optional<PlayerModel> player = playerService.getPlayerById(playerId);
+        if (player.isPresent()) {
+            PlayerModel updatedPlayer = playerService.updatePlayer(playerId, playerDetails);
+            return ResponseEntity.ok(updatedPlayer);
+        } else {
+            return ResponseEntity.notFound().build();
+        }
     }
 
     @DeleteMapping("/{id}")
-    public void deletePlayer(@PathVariable(value = "id") Long playerId) {
-        playerService.deletePlayer(playerId);
+    public ResponseEntity<Void> deletePlayer(@PathVariable(value = "id") Long playerId) {
+        Optional<PlayerModel> player = playerService.getPlayerById(playerId);
+        if (player.isPresent()) {
+            playerService.deletePlayer(playerId);
+            return ResponseEntity.noContent().build();
+        } else {
+            return ResponseEntity.notFound().build();
+        }
     }
 }
